refactor(release): separate release version policy
