+++ scripts/lib/npm-publish-plan.d.mts
@@ -1,20 +1,3 @@
-export type ParsedReleaseVersion = {
-  version: string;
-  baseVersion: string;
-  channel: "stable" | "alpha" | "beta";
-  year: number;
-  month: number;
-  patch: number;
-  alphaNumber?: number;
-  betaNumber?: number;
-  correctionNumber?: number;
-};
-export type ReleaseTrain =
-  | "alpha"
-  | "beta"
-  | "stable"
-  | "extended-stable"
-  | "unsupported-extended-stable-correction";
 export type NpmPublishPlan = {
   channel: "stable" | "alpha" | "beta";
   publishTag: "latest" | "alpha" | "beta" | "extended-stable";
@@ -35,12 +18,6 @@ export function fetchNpmRegistryPackumentWithRetry(params: {
   sleep?: (delayMs: number) => Promise<void>;
   createSignal?: (timeoutMs: number) => AbortSignal;
 }): Promise<NpmRegistryPackumentResult>;
-export function parseReleaseVersion(version: string): ParsedReleaseVersion | null;
-export function classifyReleaseTrain(parsedVersion: ParsedReleaseVersion): ReleaseTrain;
-export function collectReleaseVersionFloorErrors(
-  version: string | ParsedReleaseVersion | null,
-): string[];
-export function compareReleaseVersions(left: string, right: string): number | null;
 export function resolveNpmPublishPlan(
   version: string,
   currentBetaVersion?: string | null,
