+++ test/npm-publish-plan.test.ts
@@ -1,32 +1,12 @@
 // npm publish plan tests validate package publish planning rules.
 import { describe, expect, it } from "vitest";
 import {
-  classifyReleaseTrain,
-  collectReleaseVersionFloorErrors,
   fetchNpmRegistryPackumentWithRetry,
-  parseReleaseVersion,
   resolveNpmDistTagMirrorAuth,
   resolveNpmPublishPlan,
   resolvePublishedNpmVersionRoute,
   shouldRequireNpmDistTagMirrorAuth,
 } from "../scripts/lib/npm-publish-plan.mjs";
-
-describe("release train classification", () => {
-  it.each([
-    ["2026.7.2-alpha.1", "alpha"],
-    ["2026.7.2-beta.1", "beta"],
-    ["2026.7.32", "stable"],
-    ["2026.6.33", "extended-stable"],
-    ["2026.6.34", "extended-stable"],
-    ["2026.6.33-1", "unsupported-extended-stable-correction"],
-  ] as const)("classifies %s as %s", (version, expected) => {
-    const parsed = parseReleaseVersion(version);
-    if (!parsed) {
-      throw new Error(`test version did not parse: ${version}`);
-    }
-    expect(classifyReleaseTrain(parsed)).toBe(expected);
-  });
-});
 
 function registryResponse(params: {
   status?: number;
@@ -206,23 +186,6 @@ describe("fetchNpmRegistryPackumentWithRetry", () => {
     expect(result).toEqual({ status: 404, ok: false, packument: null });
     expect(fetchCalls).toBe(1);
     expect(cancelCalls).toBe(1);
-  });
-});
-
-describe("collectReleaseVersionFloorErrors", () => {
-  it("blocks June 2026 stable and beta release trains below the published beta floor", () => {
-    expect(collectReleaseVersionFloorErrors("2026.6.4")).toEqual([
-      'June 2026 stable and beta release trains must use patch 5 or higher because 2026.6.5-beta.1 is already published; found "2026.6.4".',
-    ]);
-    expect(collectReleaseVersionFloorErrors("2026.6.4-beta.1")).toEqual([
-      'June 2026 stable and beta release trains must use patch 5 or higher because 2026.6.5-beta.1 is already published; found "2026.6.4-beta.1".',
-    ]);
-  });
-
-  it("keeps alpha compatibility and patch-floor release trains valid during the transition", () => {
-    expect(collectReleaseVersionFloorErrors("2026.6.4-alpha.1")).toEqual([]);
-    expect(collectReleaseVersionFloorErrors("2026.6.5-beta.2")).toEqual([]);
-    expect(collectReleaseVersionFloorErrors("2026.7.1")).toEqual([]);
   });
 });
